Do not spawn new processes in unit tests
As it makes coverage tracking more complicated
(https://coverage.readthedocs.io/en/latest/subprocess.html).

Instead we invoke the main() function directly and patch sys.argv
appropriately. For tests involving IPython, we need to make sure that
the IPY_TEST_SIMPLE_PROMPT env variable is set, which we do at the very
beginning of test.py (before anything else is imported). This way we do
not require devs to set it when running the test (which granted we
could do in a wrapper script).