Add support for secure TLS channel to P4Runtime server
Until now we have been using an insecure channel, with no option to
switch to SSL/TLS.
With this change, a new flag (--ssl) is available to connect to
P4Runtime over SSL/TLS. Another flag (--cacert) is also introduced so
that users can provide the CA cert used to verify the server against
(instead of the default root certificates).
For now, backwards compatibility is preserved: by default, an insecure
channel is used.

See #84